pyproject.toml
[tool.poetry]
name = "OPTimal-summary"
version = "0.1.0"
description = ""
authors = ["Théophane", "Maud", "Marc"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
tqdm = "^4.65.0"
numpy = "^1.24.2"
torch = "^2.0.0"
hydra-core = "^1.3.2"
transformers = {extras = ["torch"], version = "^4.28.1"}
pandas = "^2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
